requirements.txt
streamlit>=1.30.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
Pillow>=10.0.0
